requirements.txt
google-generativeai
requests
python-dotenv
pydantic-ai[mcp]
fastapi
uvicorn